--- modules/react-native-maps-v2/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Maps-v2",
-  "description": "Add interactive maps to your mobile app to make it easier for users to find your locations.",
-  "root": "/modules/maps-v2"
-}
