+++ modules/react-native-maps/README.md
@@ -1,23 +1,7 @@
 # Maps
-
-## Setup
-
-Update the map's initial location in:
-`modules/maps/index.js`
-
-This is the default value (San Francisco):
-
-```javascript
-const region = {
-  latitude: 37.78825,
-  longitude: -122.4324,
-  latitudeDelta: 0.0922,
-  longitudeDelta: 0.0421,
-};
-```
+Maps module is a react-native based module. Helps user to search and navigate to the required locations.
 
 ## Android
-
 ### Add Google Maps Key
 
 Update `android/app/src/main/AndroidManifest.xml` with the following XML meta tags:
@@ -29,8 +13,6 @@ Update `android/app/src/main/AndroidManifest.xml` with the following XML meta ta
             android:name="com.google.android.geo.API_KEY"
             android:value="Your Google maps API Key Here"/>
 
-        <!-- You will also only need to add this uses-library tag -->
-        <uses-library android:name="org.apache.http.legacy" android:required="false"/>
     </application>
 ```
 
@@ -53,13 +35,55 @@ https://github.com/react-native-maps/react-native-maps/blob/master/docs/installa
 
 ## Manual Setup
 
-If you want to use the module directly, or in other modules, you can do so by importing it and using the following properties.
+1. If you want to use the module directly, or in other modules, you can do so by importing it and using the following properties.
 
 ```javascript
-import Maps from "@modules/maps";
+import Maps from "@modules/maps-v2";
 
 const { title, navigator } = Maps;
 ```
+
+2. You can call module directly by importing navigator without going through any routing. And pass the params to the module.
+
+```javascript
+
+import { modules } from '@modules';
+
+const Maps = modules[module_index].value.navigator;  //module_index : position of the module in modules folder
+
+<Maps origin={obj} originTitle = {''} apiKey={GOOGLE_MAPS_APIKEY} onNavigationStart={func} .../>
+
+```
+
+## Params
+
+Below is the list of all params that can be passed to the module.
+
+| Name              | Type       | Description                                                    |
+| ---------------   |:----------:|:---------------------------------------------------------------|
+| origin `required` | `object`   | The origin location `{latitude: , longitude: }`to start routing from.|
+| originTitle       | `string`   | Title of the origin             |
+| originDescription | `string`   | Description of the provided origin.                 |
+| apiKey `required` | `string`   | Your Google Maps API Key. Make sure you've enabled the Google Maps Directions API for that key using the Google API Console. |
+| enableDirections  | `boolean`  |Set `enableDirections=true` if you want to choose destination and start navigation between origin and destination.|
+| onNavigationStart | `function` | Called in case the navigation has started between origin to destination|
+| onNavigationError | `function` | Called in case the navigation has failed.           |
+| onLatLngChange | `function` | Called in case the marker has moved to new location.           |
+| getDistance (km)  | `function` | Returns the distance(as function param) between origin and destination.  |
+| getDuration (minutes)| `function` | Returns estimated time(as function param), needed to get from origin to destination|
+| markerColor       | `string`   | Custom color for the marker.                     |
+| markerImage       | `string`   | URL of the image to be displayed as marker.                     |
+| markerImageStyle  | `object`   | Set style for the marker image. e.g (`height`, `width`, `resizeMode`) etc. |
+| getDestinationAddress| `function` | Returns  `{latitude: , longitude}` of the destination.       |
+| strokeColor       | `string`   | Color for the line connecting origin and destination.          |
+| strokeWidth       | `number`   | Width of the line connecting origin and destination.                        |
+| showSearchInput   | `boolean`  | Set `showSearchInput={true}` to show search bar on map to search location. It's default value is `false`. |
+| mainContainerStyle| `object`   | Set style for the maps main container.  |
+| markedLocations   | `array` | Array of objects. Each object should have latitude, longitude, title and description for the location to be marked on the map.  |
+| onDragStart         | `function` | Callback that is called when the user initiates a drag on this marker.|
+| onDrag         | `function` | Callback called continuously as the marker is dragged.|
+| onDragEnd         | `function` | Callback that is called when a drag on this marker finishes. Returns locations longitude and latitude as function param.|
+
 
 ## Contributing
