+++ modules/react-native-maps/package.json
@@ -5,7 +5,10 @@
   "private": true,
   "main": "index.js",
   "x-dependencies": {
-    "react-native-maps": "0.27.1"
+    "react-native-maps": "0.27.1",
+    "react-native-maps-directions": "1.8.0",
+    "react-native-google-places-autocomplete": "2.4.1",
+    "prop-types": "15.8.1"
   },
   "author": "Crowdbotics",
   "license": "ISC"
