+++ modules/react-native-maps/styles.js
@@ -25,11 +25,42 @@ export const styles = StyleSheet.create({
   },
   searchbar: {
     backgroundColor: "white",
-    marginLeft: scale(10),
-    marginRight: scale(10),
-    marginTop: scaleVertical(5),
+    marginLeft: scale(15),
+    marginRight: scale(15),
+    marginTop: scaleVertical(10),
     marginBottom: scaleVertical(5),
-    borderRadius: 12,
-    borderColor: "#E5E5E5"
+    borderRadius: 8,
+    borderColor: "#E5E5E5",
+    zIndex: 9,
+    paddingHorizontal: 10
+  },
+  directionsContainer: {
+    backgroundColor: "white",
+    paddingHorizontal: 7,
+    paddingVertical: 5,
+    position: "absolute",
+    bottom: 30,
+    left: 30,
+    zIndex: 999,
+    borderRadius: 30,
+    elevation: 5
+  },
+  marker: { height: 40, width: 40, resizeMode: "contain" },
+  tooltip: { height: 100, width: 150, backgroundColor: "#F4F1F1", borderRadius: 10, padding: 0, margin: 0 }
+});
+
+export const autoCompleteStyles = StyleSheet.create({
+  textInputContainer: {
+    marginHorizontal: 15,
+    top: 10,
+    zIndex: 1000
+  },
+  textInput: {
+    height: 44,
+    color: "#5d5d5d",
+    fontSize: 16
+  },
+  predefinedPlacesDescription: {
+    color: "#1faadb"
   }
 });
